Moved my files over to my pc
