+++ nbproject/private/private.properties
@@ -1,6 +1,13 @@
 auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
+compile.on.save=true
+do.depend=false
+do.jar=true
+do.jlink=false
 # No need to modify this property unless customizing JavaFX Ant task infrastructure
 endorsed.javafx.ant.classpath=.
+javac.debug=true
+javadoc.preview=true
 javafx.run.inbrowser=<Default System Browser>
 javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\IEXPLORE.EXE
-user.properties.file=C:\\Users\\luoph\\AppData\\Roaming\\NetBeans\\10.0\\build.properties
+jlink.strip=false
+user.properties.file=C:\\Users\\liamd\\AppData\\Roaming\\NetBeans\\12.3\\build.properties
