+++ nbproject/project.properties
@@ -1,6 +1,5 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
@@ -39,10 +38,12 @@ includes=**
 jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
+    ${libs.JavaFX_15.classpath}:\
     ${javafx.classpath.extension}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
+javac.external.vm=false
 javac.modulepath=
 javac.processormodulepath=
 javac.processorpath=\
@@ -59,6 +60,7 @@ javac.test.processorpath=\
 javadoc.additionalparam=
 javadoc.author=false
 javadoc.encoding=${source.encoding}
+javadoc.html5=false
 javadoc.noindex=false
 javadoc.nonavbar=false
 javadoc.notree=false
@@ -98,6 +100,8 @@ javafx.preloader.type=none
 javafx.rebase.libs=false
 javafx.run.height=600
 javafx.run.width=800
+jlink.launcher=false
+jlink.launcher.name=AccidentProject
 # Pre-JavaFX 2.0 WebStart is deactivated in JavaFX 2.0+ projects
 jnlp.enabled=false
 # Main class for Java launcher
@@ -108,7 +112,8 @@ manifest.custom.codebase=*
 manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
-platform.active=JDK_1.8
+mkdist.disabled=false
+platform.active=default_platform
 run.classpath=\
     ${dist.jar}:\
     ${javac.classpath}:\
