+++ src/accidentproject/FallingSquare.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package integrativeproject.FallingObjectClasses;
+package accidentproject;
 
 import integrativeproject.FallingObjectClasses.FallingObject;
 import javafx.scene.shape.Rectangle;
